NAND and NANF

+++ WebGL/G3MWebGLSDK/src/org/glob3/mobile/specific/MathUtils_WebGL.java
@@ -18,7 +18,7 @@ public final class MathUtils_WebGL
 
    @Override
    public float NanF() {
-      return (float) Double.NaN;
+      return Float.NaN;
    }
